:sparkles: Ensure at least 1 target. (#90)
Resolves 2 problems:


The Windup CLI requires it be passed at least one `--target` parameter.
When analysis is performed using only rules files defined in
repository(s), this can result in the task.Data list of targets to be
empty because the UI does not know anything about the files in the git
repositories. Long term, windup needs to stop requiring the --target.
However, as a work-around, the addon will inject an empty rule file (no
rules) with a target of `Empty` and list it as `--target Empty` to make
windup happy.


A RuleSet (defined in rules files) may define a `targetTechnology` but
it's not required. When a target is defined, it must match a `--target`
for the rules to be executed. When a target is not defined, the rules
are executed unconditionally.
For rule files in a repository, the addon need to parse the rule file
and add each of the targets (defined in the ruleset metadata) to the
`--target` parameters.

https://issues.redhat.com/browse/MTA-304

Signed-off-by: Jeff Ortel <jortel@redhat.com>